#17: refactoring, restructured redux implementation
