--- src/redux/ActionsTypes.js
@@ -1,3 +1,0 @@
-'use strict';
-
-export const AddListItem = 'AddListItem';
