+++ src/redux/actions/shop/list.js
@@ -1,10 +1,10 @@
 'use strict';
 
-import {AddListItem} from './ActionsTypes';
+import {SHOP_ADD_LIST_ITEM} from '../../types';
 
 export function addListItem(content){
     return {
-        type: AddListItem,
+        type: SHOP_ADD_LIST_ITEM,
         payload: {
             name: content
         } // TODO: add a model here doing the transformation
