+++ src/redux/reducers/selectors.js
@@ -1,1 +1,1 @@
-export const getList = store => store.listItems;
+export const getList = store => store.listItems; // TODO: where to put this?
