+++ src/redux/store.js
@@ -1,4 +1,6 @@
 import { createStore } from "redux";
 import rootReducer from './reducers';
 
-export default createStore(rootReducer);
+const initialState = {};
+
+export default createStore(rootReducer, initialState);
