+++ src/redux/types.js
@@ -1,0 +1,3 @@
+'use strict';
+
+export const SHOP_ADD_LIST_ITEM = 'SHOP_ADD_LIST_ITEM';
